fix image uploading
fix image uploading

+++ app/controllers/product.js
@@ -64,26 +64,30 @@ module.exports.create = function(req, res){
 	}
 
 	if(image1 && image2 && image3 && image4){
-		formData.name = req.body.name;
-		formData.price = req.body.price;
-		formData.quantity = req.body.quantity;
-		formData.desc = req.body.desc;
-		formData.user = req.user;
-		formData.userName = req.user.name;
-		formData.userMobilePhone = req.user.mobilePhone;
-		formData.firstName = req.user.firstName;
-		formData.lastName = req.user.lastName;
-		formData.email = req.user.email;
+		if(!image1.truncated && !image2.truncated && !image3.truncated && !image4.truncated){
+			formData.name = req.body.name;
+			formData.price = req.body.price;
+			formData.quantity = req.body.quantity;
+			formData.desc = req.body.desc;
+			formData.user = req.user;
+			formData.userName = req.user.name;
+			formData.userMobilePhone = req.user.mobilePhone;
+			formData.firstName = req.user.firstName;
+			formData.lastName = req.user.lastName;
+			formData.email = req.user.email;
 
-		formData.save(function(err, product){
-			if(err){
-				res.status(500).jsonp({message: errorHandler.getErrorMessage(err)});
-			} else {
-				res.status(200).jsonp(product);
-			}
-		});
+			formData.save(function(err, product){
+				if(err){
+					res.status(500).jsonp({message: errorHandler.getErrorMessage(err)});
+				} else {
+					res.status(200).jsonp(product);
+				}
+			});
+		} else {
+			res.status(400).jsonp({message: 'Please upload images of max size of 1MB for every image'});
+		}
 	} else {
-		res.status(400).jsonp({message: 'Please provide four images of the product'});
+		res.status(400).jsonp({message: 'Please upload four images of the product'});
 	}
 }
 
@@ -105,55 +109,73 @@ module.exports.update = function(req, res){
 	var image1 = req.files.image1,
 		image2 = req.files.image2,
 		image3 = req.files.image3,
-		image4 = req.files.image4;
+		image4 = req.files.image4,
+		imageUploadError = false;
 
 	var formData = {};
 
 	if(!_.isEmpty(req.files)){
 		//if the user had another image then remove it before adding the new one
 		if(!_.isEmpty(image1)){
-			//get the new file name
-			formData.image1 = image1.name;
+			if(image1.truncated){
+				imageUploadError = true;
+			} else {
+				//get the new file name
+				formData.image1 = image1.name;
+			}
 		}
-
 		//if the user had another image then remove it before adding the new one
 		if(!_.isEmpty(image2)){
-			//get the new file name
-			formData.image2 = image2.name;
+			if(image2.truncated){
+				imageUploadError = true;
+			} else {
+				//get the new file name
+				formData.image2 = image2.name;
+			}
 		}
-
 		//if the user had another image then remove it before adding the new one
 		if(!_.isEmpty(image3)){
-			//get the new file name
-			formData.image3 = image3.name;
+			if(image3.truncated){
+				imageUploadError = true;
+			} else {
+				//get the new file name
+				formData.image3 = image3.name;
+			}
 		}
-
 		//if the user had another image then remove it before adding the new one
 		if(!_.isEmpty(image4)){
-			//get the new file name
-			formData.image4 = image4.name;
+			if(image4.truncated){
+				imageUploadError = true;
+			} else {
+				//get the new file name
+				formData.image4 = image4.name;
+			}
 		}
 	}
 
-	formData.name = req.body.name;
-	formData.price = req.body.price;
-	formData.quantity = req.body.quantity;
-	formData.desc = req.body.desc;
+	if(imageUploadError){
+		res.status(400).jsonp({message: 'Please upload images of max size of 1MB for every image'});
+	} else {
+		formData.name = req.body.name;
+		formData.price = req.body.price;
+		formData.quantity = req.body.quantity;
+		formData.desc = req.body.desc;
 
-	products.findOneAndUpdate({_id: req.params.id, user: req.user._id}, formData, function(err, product, numOfAffectedRows){
-		if(err){
-			fs.unlink('./public/uploads/' + req.user.image1); // delete the partially written file
-			fs.unlink('./public/uploads/' + req.user.image2); // delete the partially written file
-			fs.unlink('./public/uploads/' + req.user.image3); // delete the partially written file
-			fs.unlink('./public/uploads/' + req.user.image4); // delete the partially written file
+		products.findOneAndUpdate({_id: req.params.id, user: req.user._id}, formData, function(err, product, numOfAffectedRows){
+			if(err){
+				fs.unlink('./public/uploads/' + req.user.image1); // delete the partially written file
+				fs.unlink('./public/uploads/' + req.user.image2); // delete the partially written file
+				fs.unlink('./public/uploads/' + req.user.image3); // delete the partially written file
+				fs.unlink('./public/uploads/' + req.user.image4); // delete the partially written file
 
-			res.status(500).jsonp({message: errorHandler.getErrorMessage(err)});
-		} else if(product){
-			res.status(200).jsonp(product);
-		} else {
-			res.status(200).json({message: 'Failed to update document'});
-		}
-	});
+				res.status(500).jsonp({message: errorHandler.getErrorMessage(err)});
+			} else if(product){
+				res.status(200).jsonp(product);
+			} else {
+				res.status(200).json({message: 'Failed to update document'});
+			}
+		});
+	}
 }
 
 //delete only if the product creator has the same logged in user id which mean they are the same person
